refactor: [web] split up the appbar
