+++ packages/web/src/components/NavBar/styles.ts
@@ -1,0 +1,6 @@
+export const commonStyles = {
+    minHeight: { xs: '56px', sm: '64px' },
+    display: 'flex',
+    justifyContent: 'center',
+    alignItems: 'center'
+};
